organizing the project structure add Detail in Bottom Navigation
