+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\Раиль\.android\avd\Pixel_5_API_28.avd" />
+            <value value="C:\Users\Раиль\.android\avd\Pixel_5_API_29.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-10-22T06:20:41.870784100Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-12-20T10:20:39.639400500Z" />
   </component>
 </project>
