+++ app/src/main/java/com/example/newsfetcher/App.kt
@@ -5,7 +5,7 @@ import androidx.appcompat.app.AppCompatDelegate
 import com.example.newsfetcher.di.databaseModule
 import com.example.newsfetcher.di.networkModule
 import com.example.newsfetcher.feature.bookmarks.di.bookmarksModule
-import com.example.newsfetcher.feature.di.mainScreenModule
+import com.example.newsfetcher.feature.main_screen.di.mainScreenModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -20,6 +20,6 @@ class App : Application() {
         }
 
         // Для использования темной темы приложением
-        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM)
+//        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_FOLLOW_SYSTEM)
     }
 }
