+++ app/src/main/java/com/example/newsfetcher/canbedeleted/Sealed.kt
@@ -1,4 +1,4 @@
-package com.example.newsfetcher.feature.mainscreen
+package com.example.newsfetcher.feature.main_screen.mainscreen
 
 import android.util.Log
 
@@ -8,11 +8,12 @@ class Sealed {
     val enum: Enum = Enum.Two
 
     init {
-        when(clas){
+        when(clas) {
             is Hierarchy.Four -> Log.d("TAG", clas.text)
             is Hierarchy.One -> Log.d("TAG", clas.text)
             is Hierarchy.Three -> Log.d("TAG", clas.text)
             is Hierarchy.Two -> Log.d("TAG", clas.text + clas.index)
+            else -> {}
         }
         when(enum){
             Enum.One -> TODO()
@@ -26,10 +27,10 @@ class Sealed {
 }
 
 sealed class Hierarchy(){
-    data class One(val text: String):Hierarchy()
-    data class Two(val text: String, val index: Int):Hierarchy()
-    data class Three(val text: String):Hierarchy()
-    data class Four(val text: String):Hierarchy()
+    data class One(val text: String): Hierarchy()
+    data class Two(val text: String, val index: Int): Hierarchy()
+    data class Three(val text: String): Hierarchy()
+    data class Four(val text: String): Hierarchy()
 }
 
 enum class Enum(val text: String){
