+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks/data/Mapper.kt
@@ -1,7 +1,7 @@
 package com.example.newsfetcher.feature.bookmarks.data
 
 import com.example.newsfetcher.feature.bookmarks.data.local.model.BookmarkEntity
-import com.example.newsfetcher.feature.domian.ArticleModel
+import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
 
 fun BookmarkEntity.toDomain() = ArticleModel(
     author = author,
