+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks/di/bookmarks_module.kt
@@ -4,7 +4,7 @@ import com.example.newsfetcher.feature.bookmarks.data.local.BookmarksLocalSource
 import com.example.newsfetcher.feature.bookmarks.data.local.BookmarksRepository
 import com.example.newsfetcher.feature.bookmarks.data.local.BookmarksRepositoryImpl
 import com.example.newsfetcher.feature.bookmarks.domian.BookmarksInteractor
-import com.example.newsfetcher.feature.bookmarks.ui.BookmarksScreenViewModel
+import com.example.newsfetcher.feature.bookmarks.mainscreen.BookmarksScreenViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
