+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks/domian/BookmarksInteractor.kt
@@ -3,7 +3,7 @@ package com.example.newsfetcher.feature.bookmarks.domian
 import com.example.newsfetcher.base.Either
 import com.example.newsfetcher.base.attempt
 import com.example.newsfetcher.feature.bookmarks.data.local.BookmarksRepository
-import com.example.newsfetcher.feature.domian.ArticleModel
+import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
 
 class BookmarksInteractor(private val bookmarksRepository: BookmarksRepository){
     suspend fun create(model: ArticleModel) {
