+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks/mainscreen/BookmarksFragment.kt
@@ -1,9 +1,8 @@
-package com.example.newsfetcher.feature.bookmarks.ui
+package com.example.newsfetcher.feature.bookmarks.mainscreen
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.viewModelScope
 import com.example.newsfetcher.R
 import org.koin.androidx.viewmodel.ext.android.viewModel
