+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks/mainscreen/Contract.kt
@@ -1,7 +1,8 @@
-package com.example.newsfetcher.feature.bookmarks.ui
+package com.example.newsfetcher.feature.bookmarks.mainscreen
 
 import com.example.newsfetcher.base.Event
-import com.example.newsfetcher.feature.domian.ArticleModel
+import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
+
 
 data class ViewState(
     val bookmarksArticle: List<ArticleModel>
