--- app/src/main/java/com/example/newsfetcher/feature/data/ArticlesRepository.kt
@@ -1,7 +1,0 @@
-package com.example.newsfetcher.feature.data
-
-import com.example.newsfetcher.feature.domian.ArticleModel
-
-interface ArticlesRepository {
-    suspend fun getArticles() : List<ArticleModel>
-}
