--- app/src/main/java/com/example/newsfetcher/feature/data/model/ArticlesRemoteRepositoryImpl.kt
@@ -1,14 +1,0 @@
-package com.example.newsfetcher.feature.data.model
-
-import com.example.newsfetcher.feature.data.ArticlesRemoteSource
-import com.example.newsfetcher.feature.data.ArticlesRepository
-import com.example.newsfetcher.feature.data.toDomian
-import com.example.newsfetcher.feature.domian.ArticleModel
-
-class ArticlesRemoteRepositoryImpl(private val source: ArticlesRemoteSource) : ArticlesRepository {
-    override suspend fun getArticles(): List<ArticleModel> {
-        return source.getArticles().articlesList.map {
-            it.toDomian()
-        }
-    }
-}
