+++ app/src/main/java/com/example/newsfetcher/feature/main_screen/data/ArticlesRemoteSource.kt
@@ -1,6 +1,6 @@
-package com.example.newsfetcher.feature.data
+package com.example.newsfetcher.feature.main_screen.data
 
-import com.example.newsfetcher.feature.data.model.ArticlesRemoteModel
+import com.example.newsfetcher.feature.main_screen.data.model.ArticlesRemoteModel
 
 class ArticlesRemoteSource(private val api: NewsAPI) {
     suspend fun getArticles(): ArticlesRemoteModel {
