+++ app/src/main/java/com/example/newsfetcher/feature/main_screen/data/ArticlesRepository.kt
@@ -1,0 +1,7 @@
+package com.example.newsfetcher.feature.main_screen.data
+
+import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
+
+interface ArticlesRepository {
+    suspend fun getArticles() : List<ArticleModel>
+}
