+++ app/src/main/java/com/example/newsfetcher/feature/main_screen/data/Mapper.kt
@@ -1,7 +1,7 @@
-package com.example.newsfetcher.feature.data
+package com.example.newsfetcher.feature.main_screen.data
 
-import com.example.newsfetcher.feature.data.model.ArticleRemoteModel
-import com.example.newsfetcher.feature.domian.ArticleModel
+import com.example.newsfetcher.feature.main_screen.data.model.ArticleRemoteModel
+import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
 
 fun ArticleRemoteModel.toDomian() = ArticleModel(
     title = title ?: "",
