+++ app/src/main/java/com/example/newsfetcher/feature/main_screen/mainscreen/MainScreenViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.newsfetcher.feature.mainscreen
+package com.example.newsfetcher.feature.main_screen.mainscreen
 
 import android.util.Log
 import androidx.lifecycle.*
@@ -6,7 +6,7 @@ import com.example.newsfetcher.base.BaseViewModel
 import com.example.newsfetcher.base.Event
 
 import com.example.newsfetcher.feature.bookmarks.domian.BookmarksInteractor
-import com.example.newsfetcher.feature.domian.ArticlesInteractor
+import com.example.newsfetcher.feature.main_screen.domian.ArticlesInteractor
 import kotlinx.coroutines.launch
 
 class MainScreenViewModel(
