+++ app/src/main/res/menu/menu.xml
@@ -7,5 +7,8 @@
     <item android:id="@+id/itemBookmarks"
         android:icon="@drawable/ic_baseline_bookmark_border_24"
         android:title="@string/favorite"/>
+    <item android:id="@+id/itemDetails"
+        android:icon="@drawable/ic_baseline_zoom_out_map_24"
+        android:title="@string/detail"/>
 
 </menu>
