+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">NewsFetcher</string>
     <string name="main">Главная</string>
     <string name="favorite">Избранное</string>
+    <string name="detail">Подробно</string>
 </resources>
